+++ src/platform/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
 * [BC BREAK] Change `Symfony\AI\Platform\Contract\JsonSchema\Factory` constructor signature in order to make schema generation extensible
+* Generate JSON schema from Symfony Validator constraints when available
 
 0.4
 ---